+++ src/client/reducers/districtDrawing.ts
@@ -266,6 +266,7 @@ const districtDrawingReducer: LoopReducer<ProjectState, Action> = (
     case getType(toggleEvaluate):
       return {
         ...state,
+        evaluateMetric: undefined,
         evaluateMode: action.payload
       };
     case getType(selectEvaluationMetric):